Patch storage/postgres from master@8672764438
