+++ src/storage/postgres/src/main/java/org/locationtech/geogig/storage/postgresql/PGRepositoryResolver.java
@@ -21,6 +21,10 @@ import org.locationtech.geogig.repository.RepositoryResolver;
 import org.locationtech.geogig.repository.impl.GeoGIG;
 import org.locationtech.geogig.repository.impl.GlobalContextBuilder;
 import org.locationtech.geogig.storage.ConfigDatabase;
+import org.locationtech.geogig.storage.postgresql.config.Environment;
+import org.locationtech.geogig.storage.postgresql.config.EnvironmentBuilder;
+import org.locationtech.geogig.storage.postgresql.config.PGStorage;
+import org.locationtech.geogig.storage.postgresql.v9.PGConfigDatabase;
 
 import com.google.common.base.Optional;
 import com.google.common.base.Preconditions;
@@ -107,7 +111,7 @@ public class PGRepositoryResolver extends RepositoryResolver {
         try {
             config = new EnvironmentBuilder(repoURI).build();
         } catch (RuntimeException e) {
-            Throwables.propagateIfInstanceOf(e, IllegalArgumentException.class);
+            Throwables.throwIfInstanceOf(e, IllegalArgumentException.class);
             throw new IllegalArgumentException("Error parsing URI " + repoURI, e);
         }
         Preconditions.checkArgument(config.getRepositoryName() != null,
@@ -123,6 +127,11 @@ public class PGRepositoryResolver extends RepositoryResolver {
         hints.set(Hints.REPOSITORY_URL, repositoryLocation.toString());
         Context context = GlobalContextBuilder.builder().build(hints);
         Repository repository = new GeoGIG(context).getRepository();
+        // Ensure the repository exists. If it's null, we might have a non-existing repo URI location
+        if (repository == null) {
+            throw new RepositoryConnectionException("Could not connect to repository. Check that the URI is valid: " +
+            repositoryLocation);
+        }
         repository.open();
         return repository;
     }
