+++ src/storage/postgres/src/test/java/org/locationtech/geogig/storage/postgresql/PGTestDataSourceProvider.java
@@ -17,7 +17,9 @@ import org.junit.AssumptionViolatedException;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.rules.ExternalResource;
-import org.locationtech.geogig.storage.postgresql.Environment.ConnectionConfig;
+import org.locationtech.geogig.storage.postgresql.config.ConnectionConfig;
+import org.locationtech.geogig.storage.postgresql.config.Environment;
+import org.locationtech.geogig.storage.postgresql.config.PGStorage;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -32,7 +34,7 @@ public class PGTestDataSourceProvider extends ExternalResource {
 
     private static final Logger LOG = LoggerFactory.getLogger(PGTestDataSourceProvider.class);
 
-    private Environment.ConnectionConfig connectionConfig;
+    private ConnectionConfig connectionConfig;
 
     private PGTestProperties props;
