+++ src/storage/postgres/src/test/java/org/locationtech/geogig/storage/postgresql/integration/PGGraphDatabaseTest.java
@@ -14,11 +14,11 @@ import org.junit.Rule;
 import org.locationtech.geogig.repository.Platform;
 import org.locationtech.geogig.storage.GraphDatabase;
 import org.locationtech.geogig.storage.impl.GraphDatabaseTest;
-import org.locationtech.geogig.storage.postgresql.Environment;
-import org.locationtech.geogig.storage.postgresql.PGGraphDatabase;
-import org.locationtech.geogig.storage.postgresql.PGStorage;
 import org.locationtech.geogig.storage.postgresql.PGTemporaryTestConfig;
 import org.locationtech.geogig.storage.postgresql.PGTestDataSourceProvider;
+import org.locationtech.geogig.storage.postgresql.config.Environment;
+import org.locationtech.geogig.storage.postgresql.config.PGStorage;
+import org.locationtech.geogig.storage.postgresql.v9.PGGraphDatabase;
 
 public class PGGraphDatabaseTest extends GraphDatabaseTest {
